requirements.txt
pyparsing
selenium
2captcha-python
twilio
python-pushover
undetected_chromedriver
selenium-wire
